implement logic of mission progress handling

+++ MA.RewardService.Domain/Abstractions/IPointsCalculationService.cs
@@ -2,5 +2,5 @@ namespace MA.RewardService.Domain.Abstractions;
 
 public interface IPointsCalculator
 {
-    int Calculate(int[] spinResult);
+    int Calculate(byte[] spinResult);
 }
